test/test_output/workspace_5-3/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-magic==0.4.27
pyclamd==1.0.2
redis==4.5.5
python-jose==3.3.0
passlib==1.7.4